+++ bin/vcp.ps1
@@ -1,0 +1,2 @@
+#!/usr/bin/env pwsh
+py -m vcp_cli @args