+++ scripts/check-version-consistency.sh
@@ -74,6 +74,7 @@ stale_versions=(
   "v0.4.2"
   "v0.4.3"
   "v0.4.4"
+  "v0.5.0"
 )
 
 entry_files=(